Interactable & OpenDoor Script
This is the submission with the interactable and open door script. I made the interactable script based off the player to try and reduce the amount of scripts we need to go through for debugging. The open door script is done, but we need to discuss how we want them to open.